accessible/tests/mochitest/test_elm_plugin.html

<!DOCTYPE html>
<html>
<head>
  <title>Plugin tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="common.js"></script>
  <script type="application/javascript"
          src="role.js"></script>
  <script type="application/javascript"
          src="states.js"></script>

  <script type="application/javascript">

    function doTest()
    {
      if (!WIN) {
        ok(true,
           "Nothing to test because accessible plugins are supported on Windows only");

        SimpleTest.finish();
        return;
      }

      testRole("plugin", ROLE_EMBEDDED_OBJECT);
      testStates("plugin", STATE_UNAVAILABLE);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="embed and object HTML tags should be given an accessible role of embedded object"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=485270">Mozilla Bug 485270</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <embed id="plugin" type="application/x-test" width="300" height="300"></embed>
</body>
</html>
